requirements.txt
gpt4all
python-dotenv